fix animes wrong display - add confirmation when delete favourite

+++ frontend/src/views/Favourites.vue
@@ -24,9 +24,7 @@
                 xmlns="http://www.w3.org/2000/svg"
                 width="18"
                 height="18"
-                :fill="
-                  anime.anime_details.is_favourite ? 'red' : 'currentColor'
-                "
+                fill="red"
                 class="bi bi-heart-fill"
                 viewBox="0 0 16 16"
               >
@@ -68,8 +66,10 @@ export default {
 
   methods: {
     async deleteFavourite(anime_id) {
-      await axios.delete(`api-v1/animes-favourites/${anime_id}/`);
-      this.setAnimesData();
+      if (confirm("Do you really want to delete this favourite ?")) {
+        await axios.delete(`api-v1/animes-favourites/${anime_id}/`);
+        this.setAnimesData();
+      }
     },
 
     async setAnimesData() {
